feat: Completed dashboard
